add comment, add status chat, add props onLongPress on message
