+++ App/Themes/Images.js
@@ -12,7 +12,11 @@ const images = {
   emptyRoom: require('../Images/empty-room.png'),
   emptyChat: require('../Images/empty-chat.png'),
   attach: require('../Images/attach.png'),
-  send: require('../Images/send.png')
+  send: require('../Images/send.png'),
+  isRead: require('../Images/isRead.png'),
+  isDelivered: require('../Images/isDelivered.png'),
+  isPending: require('../Images/isPending.png'),
+  isFailed: require('../Images/isFailed.png')
 }
 
 export default images
